Remove most unsafety in big int.
The safety guarantees that aren't validated immediately have been
removed (so only keeping the unsafe code in the hi bit extraction, which
is checked right in the call), and this leads to ~3% degradation in
performance in the worst cases, which is still 90% faster than core so
it's well worth it. The unsafe code in the multiplication algorithms is
the cause of the slowdown, but well worth it.