Remove all remaining potential unsoundness in the decimal writer.
This in general for simple cases due to slight refactoring improves performance, however, it does have a noticeable (+2-7%) penalty in some corner cases which invoke the close-to-halfway rounding cases. That said, performance of this implementation of the Dragonbox algorithm without any unsafety always outperforms Ryu, which means it's still the fastest implementation out there even with slight regressions in corner cases.