Fix issues with writing radix numbers.
This uses an unoptimized approach for non-decimal numbers, which calculates the number of digits rather than write the digits and shift them into place. Due to the high overhead of memcpy, this is likely to be more performant, however, this has noticeable overhead issues with power-of-two and 128-bit integers.

This is because 128-bit division is very slow and should be optimized. Likewise, power-of-two can use the always correct log2 approaches to get the number of digits. Non-power-of-two and non-decimal digits might have performance issues, due to the rounding issues of `log` implementations, safety is more important due to the possibility of corner cases causing issues.