Major revamp to the code structure for better build times and MSRV guarantees.
This lowers the required MSRV to 1.61.0 and ensures that all tests with dependencies, as well as the ASM, binary size, and benchmark crates are moved to a dedicated submodule. This keeps all the core code the same, in fact, no source code was changed, however, it simplifies the builds and other tests.